# api/public/.well-known/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Arch Tools API",
    "version": "1.11.0",
    "description": "64 AI-powered tools in a single API — web scraping, AI generation, crypto data, NLP, research, image generation, and more. Pay per use with API key or x402 USDC micropayments. Full platform at https://archtools.dev",
    "contact": {
      "name": "Arch Tools Support",
      "url": "https://archtools.dev/developers",
      "email": "support@archtools.dev"
    },
    "license": {
      "name": "Proprietary",
      "url": "https://archtools.dev/terms"
    }
  },
  "servers": [
    {
      "url": "https://archtools.dev",
      "description": "Production"
    }
  ],
...
  "paths": {
...
    "/v1/tools/summarize": {
      "post": {
        "operationId": "summarize",
        "summary": "Summarize text in 5 styles (bullets, tldr, executive…)",
        "tags": ["AI"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["text"],
                "properties": {
                  "text": { "type": "string" },
                  "style": { "type": "string", "enum": ["bullets", "tldr", "executive", "paragraph", "headline"], "default": "bullets" },
                  "max_length": { "type": "integer" }
                }
              }
            }
          }
        },
        "responses": { "200": { "description": "Summarized text" } }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key",
        "description": "API key obtained from registration. Pass in x-api-key header."
      },
      "x402Payment": {
        "type": "apiKey",
        "in": "header",
        "name": "X-PAYMENT",
        "description": "x402 payment header for USDC micropayments. Alternative to API key auth."
      },
      "OAuth2": {
        "type": "oauth2",
        "description": "OAuth 2.1 with PKCE",
        "flows": {
          "authorizationCode": {
            "authorizationUrl": "https://archtools.dev/oauth/authorize",
            "tokenUrl": "https://archtools.dev/oauth/token",
            "scopes": {
              "tools:read": "Access tool endpoints",
              "tools:write": "Execute tool operations",
              "agent:read": "Read agent profile",
              "agent:write": "Modify agent profile"
            }
          }
        }
      }
    },
...
  },
  "security": [
    { "ApiKeyAuth": [] },
    { "x402Payment": [] }
  ]
}